refactor: Apply Sonar lint

+++ platform/src/main/java/com/neoga/platform/bid/config/SchelduleConfig.java
@@ -54,7 +54,7 @@ public class SchelduleConfig {
         SimpleTriggerFactoryBean trigger = new SimpleTriggerFactoryBean();
         trigger.setJobDetail(job);
 
-        int frequencyInSec = 60;
+        long frequencyInSec = 60;
         log.info("Configuring trigger to fire every {} seconds", frequencyInSec);
 
         trigger.setRepeatInterval(frequencyInSec * 1000);
